Pipeline_data_analytics/Pipeline_data_analytics/kafkaProducer/requirements.txt
cassandra-driver==3.28.0
click==8.1.7
geomet==0.2.1.post1
kafka==1.3.5
kafka-python==2.0.2
six==1.16.0
